fix input and output filter on perfoms sidebar
